+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 - fix a bug where the terminal would not close when the network runs out of power
 - only draw power from the network via quantum bridge until the network's power buffer is at least half full
+- reverse-charge the network if it is below 50% (unless the terminal is also below 50%)